+++ demo/main.py
@@ -6,10 +6,9 @@ def render(item:'Todo'):
     return Li(item.title, dellink, id=id)
 
 auth = user_pwd_auth(user='s3kret', skip=[r'/favicon\.ico', r'/static/.*', r'.*\.css'])
-app,todos,Todo = fast_app(
+app,rt,todos,Todo = fast_app(
     'data/tbl.db', middleware=[auth], render=render,
     id=int, title=str, pk='id')
-rt = app.route
 
 @rt("/")
 async def get(request):