+++ fasthtml/__init__.py
@@ -1,6 +1,7 @@
-__version__ = "0.1.1"
+__version__ = "0.1.7"
 from .core import *
 from .authmw import *
 from .components import *
 from .xtend import *
 from .live_reload import *
+from .toaster import *