+++ fasthtml/common.py
@@ -9,4 +9,3 @@ from fastlite import *
 from . import *
 from .js import *
 from .fastapp import *
-