+++ fasthtml/xt.py
@@ -1,0 +1,4 @@
+from fastcore.xml import *
+from .components import *
+from .xtend import *
+